Refactor main function, breaks out process management, signal handling, process lifecycle and argument parsing modules to own files
