+++ server/src/main.c
@@ -1,436 +1,46 @@
+/**
+ * @file main.c
+ * @brief Energy Saver Deluxe - Main entry point
+ * 
+ * This application manages multiple child processes:
+ * - HTTP Server: Serves web requests
+ * - Fetchers: Retrieve energy price and weather data
+ * - Energy Advisor: Analyzes data and generates advice
+ * 
+ * All process lifecycle management is delegated to dedicated modules.
+ */
+
 #define _POSIX_C_SOURCE 200809L
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <signal.h>
-#include <errno.h>
 #include <string.h>
-#include <time.h>
+#include <unistd.h>
 
-#include "http/server/http_server.h"
-#include "logger/logger.h"
-#include "routes/root_handler.h"
 #include "cli/cli.h"
 #include "process_manager/process_manager.h"
-#include "config/config.h"
-#include "file_helper/file_helper.h"
-#include "energy_advisor/energy_advisor.h"
+#include "parent_supervisor.h"
+#include "process_lifecycle.h"
 
-#define ENERGY_ADVISOR_CHECK_INTERVAL_SECONDS 60
-
-static volatile sig_atomic_t http_server_should_quit = 0;
-static volatile sig_atomic_t process_manager_should_quit = 0;
-static volatile sig_atomic_t energy_advisor_should_quit = 0;
-static volatile sig_atomic_t *active_quit_flag = NULL;
-
-typedef struct
+/**
+ * @brief CLI help callback
+ */
+static void help_callback(void)
 {
-    Logger *logger;
-} HTTP_Cool_Context;
-
-static void install_signal_handler(int signal, void (*handler)(int))
-{
-    struct sigaction sa;
-    memset(&sa, 0, sizeof(sa));
-    sa.sa_handler = handler;
-    sigemptyset(&sa.sa_mask);
-    sa.sa_flags = 0;  // Don't use SA_RESTART - allow interruption of system calls
-
-    sigaction(signal, &sa, NULL);
-}
-
-static void generic_quit_handler(int signal)
-{
-    (void)signal;
-    if (active_quit_flag != NULL)
-        *active_quit_flag = 1;
-}
-
-void setup_http_server_signals(void)
-{
-    active_quit_flag = &http_server_should_quit;
-    install_signal_handler(SIGQUIT, generic_quit_handler);
-    install_signal_handler(SIGTERM, generic_quit_handler);
-}
-
-void setup_energy_advisor_signals(void)
-{
-    active_quit_flag = &energy_advisor_should_quit;
-    install_signal_handler(SIGQUIT, generic_quit_handler);
-    install_signal_handler(SIGTERM, generic_quit_handler);
-}
-
-void setup_process_manager_signals(void)
-{
-    active_quit_flag = &process_manager_should_quit;
-    install_signal_handler(SIGTERM, generic_quit_handler);
-    install_signal_handler(SIGINT, generic_quit_handler);
-}
-
-void help_callback(void)
-{
-    printf("Current arguments:\n--port(-h) <integer>\n--help(-h) for information\n--test(t) <string> to print self out\n");
+    printf("Energy Saver Deluxe\n");
+    printf("Usage: EnergySaverDeluxe [options]\n\n");
+    printf("Options:\n");
+    printf("  --port, -p <integer>  Server port (not yet implemented)\n");
+    printf("  --test, -t <string>   Test string (not yet implemented)\n");
+    printf("  --help, -h            Show this help message\n");
     exit(0);
 }
 
-static int parse_command_args(const char *cmd_string, char ***args_out)
-{
-    if (!cmd_string || !args_out)
-        return 0;
-
-    // Copy string since we'll modify it
-    char *str = strdup(cmd_string);
-    if (!str)
-        return 0;
-
-    // Count args first (rough estimate)
-    int max_args = 1; // Start with program name
-    for (const char *p = cmd_string; *p; p++)
-    {
-        if (*p == ' ' || *p == '\t')
-            max_args++;
-    }
-    max_args += 2; // Extra space + NULL terminator
-
-    char **args = calloc(max_args, sizeof(char *));
-    if (!args)
-    {
-        free(str);
-        return 0;
-    }
-
-    int argc = 0;
-    char *p = str;
-
-    while (*p)
-    {
-        // Skip whitespace
-        while (*p == ' ' || *p == '\t')
-            p++;
-        if (!*p)
-            break;
-
-        char *start = p;
-
-        // Handle quoted strings
-        if (*p == '\'' || *p == '"')
-        {
-            char quote = *p++;
-            start = p;
-            while (*p && *p != quote)
-                p++;
-            if (*p)
-            {
-                *p++ = '\0';
-            }
-        }
-        else
-        {
-            // Regular token
-            while (*p && *p != ' ' && *p != '\t')
-                p++;
-            if (*p)
-                *p++ = '\0';
-        }
-
-        args[argc++] = strdup(start);
-    }
-
-    args[argc] = NULL;
-    free(str);
-
-    *args_out = args;
-    return argc;
-}
-
-// Function to free the args array
-static void free_args(char **args)
-{
-    if (!args)
-        return;
-    for (int i = 0; args[i]; i++)
-    {
-        free(args[i]);
-    }
-    free(args);
-}
-
-// HTTP Server process entry point
-int http_server_process(void *context)
-{
-    (void)context;
-
-    Logger logger = {0};
-    Logger_Init(&logger, "HTTP-Server", NULL, NULL, LOGGER_OUTPUT_TYPE_CONSOLE);
-
-    setup_http_server_signals();
-
-    HTTP_Server http_server;
-
-    if (HTTP_Server_Initialize(&http_server, 1024, NULL) == false)
-    {
-        Logger_Write(&logger, "Server failed to initialize");
-        return 1;
-    }
-
-    HTTP_Server_Register_Route(&http_server, "/", HTTP_METHOD_GET, root_handler_handle);
-
-    if (HTTP_Server_Start(&http_server, 8080) == false)
-    {
-        Logger_Write(&logger, "Server failed to start");
-        return 2;
-    }
-
-    while (http_server_should_quit == 0)
-    {
-        HTTP_Server_Work(&http_server);
-    }
-
-    Logger_Write(&logger, "Disposing HTTP Server");
-    HTTP_Server_Dispose(&http_server);
-
-    Logger_Dispose(&logger);
-    return 0;
-}
-
-int energy_advisor_process(void *context)
-{
-    (void)context;
-
-    setup_energy_advisor_signals();
-
-    Config_t *cfg = Config_Get_Instance(NULL);
-
-    size_t fetcher_command_count = Config_Get_Field_Value_Integer(cfg, "fetchers_commands_count", NULL);
-    char **args = NULL;
-
-    bool first_file_exists = false;
-    bool second_file_exists = false;
-
-    while (energy_advisor_should_quit == 0)
-    {
-        for (size_t i = 0; i < fetcher_command_count; i++)
-        {
-            char *directory;
-            char *filename;
-            char *cmd_args_string = Config_Get_Field_Value_From_String_Array(cfg, "fetchers_commands_args", i);
-            parse_command_args(cmd_args_string, &args);
-
-            for (int j = 0; args[j]; j++)
-            {
-                if (strcmp(args[j], "-o") == 0 && args[j + 1])
-                {
-                    directory = args[j + 1];
-                }
-
-                if (strcmp(args[j], "-n") == 0 && args[j + 1])
-                {
-                    filename = args[j + 1];
-                }
-            }
-
-            char full_path[128];
-            snprintf(full_path, sizeof(full_path), "%s/%s", directory, filename);
-
-            if (File_Helper_File_Exists(full_path))
-            {
-                if (i == 0)
-                {
-                    first_file_exists = true;
-                }
-                else
-                {
-                    second_file_exists = true;
-                }
-            }
-
-            if (args != NULL)
-                free_args(args);
-        }
-
-        if (first_file_exists == true && second_file_exists == true)
-        {
-            Energy_Status status = Energy_Advisor_Advice();
-            if (status != ENERGY_STATUS_OK)
-            {
-                printf("Energy Advice data is missing, error code: %d\n", status);
-                return -1;
-            }
-        }
-
-        first_file_exists = false;
-        second_file_exists = false;
-
-        sleep(ENERGY_ADVISOR_CHECK_INTERVAL_SECONDS);
-    }
-
-    return 0;
-}
-
-static int run_parent_loop(pid_t process_manager_pid)
-{
-    char buffer[32] = {0};
-    bool should_quit = false;
-    while (should_quit == false)
-    {
-        if (fgets(buffer, sizeof(buffer), stdin) == NULL)
-            continue;
-        if (strncmp(buffer, "q", 1) == 0)
-        {
-            // Send SIGTERM to child process - let child handle cleanup
-            if (kill(process_manager_pid, SIGTERM) == -1)
-            {
-                printf("Error: Did not correctly kill the process: %d\n", errno);
-                return -4;
-            }
-
-            int stat_loc;
-            waitpid(process_manager_pid, &stat_loc, 0);
-            printf("stat loc: %d\n", stat_loc);
-            should_quit = true;
-        }
-    }
-    printf("Goodbye, process done.\n");
-    return 0;
-}
-
-static int update_fetcher_args_date(char **args)
-{
-    for (int j = 0; args != NULL && args[j] != NULL; j++)
-    {
-        if (strcmp(args[j], "-u") == 0 && args[j + 1])
-        {
-            if (strcmp(args[j + 1], "https://www.elprisetjustnu.se") != 0)
-            {
-                return 0;
-            }
-        }
-        if (strcmp(args[j], "-r") == 0 && args[j + 1])
-        {
-            time_t t = time(NULL);
-            struct tm tm = *localtime(&t);
-            char date_buffer[64];
-            snprintf(date_buffer, sizeof(date_buffer), "/api/v1/prices/%04d/%02d-%02d_SE4.json", tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday);
-            free(args[j + 1]);
-            args[j + 1] = strdup(date_buffer);
-        }
-    }
-
-    return 0;
-}
-
-static int run_process_manager_child(ProcessManager *process_manager)
-{
-    setup_process_manager_signals();
-
-    Logger process_manager_logger = {0};
-    if (Logger_Init(&process_manager_logger, "Process Manager", NULL, NULL, LOGGER_OUTPUT_TYPE_CONSOLE) != LOGGER_RESULT_OK)
-    {
-        printf("Failed to initialize logger for Process Manager\n");
-        return -1;
-    }
-    Logger_Write(&process_manager_logger, "%s", "Process Manager started");
-
-    if (!ProcessManager_Init(process_manager, &process_manager_logger))
-    {
-        Logger_Write(&process_manager_logger, "Failed to initialize Process Manager");
-        return -1;
-    }
-
-    pid_t server_pid = ProcessManager_Spawn(process_manager, "HTTP Server", http_server_process, NULL, false);
-
-    if (server_pid < 0)
-    {
-        Logger_Write(&process_manager_logger, "Failed to spawn HTTP Server process");
-        return -1;
-    }
-
-    Config_t *cfg = Config_Get_Instance("settings.json");
-    if (cfg == NULL)
-    {
-        Logger_Write(&process_manager_logger, "Failed to load configuration!");
-        exit(-1);
-    }
-
-    char *fetcher_exec_path = Config_Get_Field_Value_String(cfg, "fetcher_exec_path");
-    size_t fetcher_command_count = Config_Get_Field_Value_Integer(cfg, "fetchers_commands_count", NULL);
-    pid_t fetcher_pid[fetcher_command_count];
-    char **args = NULL;
-
-    for (size_t i = 0; i < fetcher_command_count; i++)
-    {
-        char *cmd_args_string = Config_Get_Field_Value_From_String_Array(cfg, "fetchers_commands_args", i);
-        parse_command_args(cmd_args_string, &args);
-
-        update_fetcher_args_date(args);
-
-        fetcher_pid[i] = ProcessManager_SpawnByExecutable(process_manager, "Fetcher", fetcher_exec_path, args, true);
-
-        if (fetcher_pid[i] < 0)
-        {
-            Logger_Write(&process_manager_logger, "Failed to spawn fetcher process");
-            return -1;
-        }
-
-        if (args != NULL)
-            free_args(args);
-    }
-
-    pid_t energy_advisor_pid = ProcessManager_Spawn(process_manager, "Energy Advisor", energy_advisor_process, NULL, false);
-
-    if (energy_advisor_pid < 0)
-    {
-        Logger_Write(&process_manager_logger, "Failed to spawn Energy Advisor process");
-        return -1;
-    }
-
-    struct timespec ts;
-    ts.tv_sec = 1;
-    ts.tv_nsec = 0;
-
-    // Wait for child processes to finish or termination signal
-    while (process_manager_should_quit == 0)
-    {
-        for (size_t fetcher_index = 0; fetcher_index < fetcher_command_count; fetcher_index++)
-        {
-            pid_t pid = fetcher_pid[fetcher_index];
-            char buffer[256] = {0};
-            ssize_t bytes_read = ProcessManager_ReadFromChild(process_manager, pid, buffer, sizeof(buffer) - 1);
-            if (bytes_read > 0)
-            {
-                buffer[bytes_read] = '\0';
-                Logger_Write(&process_manager_logger, "Output from fetcher process: %s", buffer);
-                if (strcmp(buffer, "NEW_DATA") == 0)
-                {
-                    Logger_Write(&process_manager_logger, "Received NEW_DATA from fetcher process");
-                    ProcessManager_WriteToChild(process_manager, pid, "ACK", 4);
-                }
-            }
-        }
-        nanosleep(&ts, NULL);
-    }
-
-    Logger_Write(&process_manager_logger, "Process Manager shutting down...");
-    ProcessManager_TerminateAll(process_manager);
-
-    // Wait for all child processes to terminate
-    Logger_Write(&process_manager_logger, "Waiting for child processes to exit...");
-    int status;
-    while (wait(&status) > 0)
-    {
-        // Reap all child processes
-    }
-    Config_Instance_Dispose();
-    ProcessManager_Destroy(process_manager);
-    Logger_Dispose(&process_manager_logger);
-    return 0;
-}
+// TODO Add restart of processes when api error etc, logger macro?
 
 int main(int argc, char **argv)
 {
+    // Parse command line arguments
     CLI cli;
     int port_argument_data = 0;
     char test_string[128];
@@ -446,14 +56,24 @@ int main(int argc, char **argv)
         return -1;
     }
 
+    // Fork to create parent supervisor and process manager child
     ProcessManager process_manager;
     pid_t process_manager_pid = fork();
 
     if (process_manager_pid > 0)
+    {
+        // Parent process: supervise the process manager
         return run_parent_loop(process_manager_pid);
-
-    if (process_manager_pid == 0)
+    }
+    else if (process_manager_pid == 0)
+    {
+        // Child process: run the process manager
         return run_process_manager_child(&process_manager);
-
-    return 0;
+    }
+    else
+    {
+        // Fork failed
+        perror("fork");
+        return -1;
+    }
 }
